+++ R/coxphmulti.R
@@ -31,10 +31,11 @@
 #' 	fit2df()
 
 coxphmulti <- function(.data, dependent, explanatory){
+  requireNamespace("survival")
   result = list()
   for (i in 1:length(dependent)){
-    result[[i]] = survival::coxph(as.formula(paste0("survival::", dependent, "~",
-                                                    paste(explanatory, collapse="+"))), data=.data)
+    result[[i]] = coxph(as.formula(paste0(dependent, "~",
+                                          paste(explanatory, collapse="+"))), data=.data)
   }
   result = setNames(result, dependent)
   class(result) = "coxphlist"