+++ tests/testthat/test_plots.R
@@ -9,6 +9,16 @@ test_that("hr_plot gives plot", {
 	expect_is(hr_plot(colon_s, "Surv(time, status)", "age.factor"), "grob")
 })
 
+
+test_that("or_plot gives plot with remove ref level", {
+	expect_is(or_plot(colon_s, "mort_5yr", "age.factor", remove_ref = TRUE), "grob")
+})
+
+test_that("hr_plot gives plot with remove ref level", {
+	expect_is(hr_plot(colon_s, "Surv(time, status)", "age.factor", remove_ref = TRUE), "grob")
+})
+
+
 test_that("surv_plot gives plot", {
 	expect_is(surv_plot(colon_s, "Surv(time, status)", "age.factor"), "ggsurvplot")
 })